# Use np.interp instead of interpolation.interp