This PR replaces `interpolation.interp` to `np.interp` now that `numba` supports `np.interp` directly.  Also, removes `plt.rcParams["figure.figsize"] = (11, 5)  #set default figure size` redundant line.