@jstac 2024-03-06
Many thanks @kp992 .

@mmcky or @HumphreyYang , could you please do a first round review of this?

The aim is that swapping to NumPy's interp should not change outputs or degrade performance.