@jstac 2024-03-06
Note that this is related to https://github.com/QuantEcon/lecture-python-advanced.myst/pull/156