@mmcky 2024-03-06
@jstac re: performance there are small degradations 

| lecture | current time | new time |
|--------|-------------|-----------|
| cake_eating_numerical | 30.4 | 33.8 |