@mmcky 2024-03-06
Remaining lectures to migrate

- [ ] lectures/jv.md:from interpolation import interp
- [ ] lectures/mccall_correlated.md:from interpolation import interp
- [ ] lectures/mccall_fitted_vfi.md:from interpolation import interp
- [ ] lectures/navy_captain.md:from interpolation import interp
- [ ] lectures/odu.md:from interpolation import mlinterp, interp
- [ ] lectures/optgrowth_fast.md:from interpolation import interp
- [ ] lectures/wald_friedman.md:from interpolation import interp